Добавлена библиотека typed-redux-saga для типизации Redux-Saga. Полностью типизирован Redux
