+++ src/store/root.saga.ts
@@ -1,12 +1,12 @@
 import {
 	all,
 	call
-} from 'redux-saga/effects'
+} from 'typed-redux-saga/macro' // 'redux-saga/effects'
 import categoriesSaga from "./categories/category.saga";
 import usersSagas from "./user/user.saga";
 
 export function* rootSaga() {
-	yield all([
+	yield* all([
 		call(categoriesSaga),
 		call(usersSagas),
 	])
